+++ KmtBackend.Models/DTOs/Leave/UpdateLeaveBalanceRequest.cs
@@ -3,6 +3,6 @@
     public class UpdateLeaveBalanceRequest
     {
         public int? TotalDays { get; set; } = null;
-        public int? UsedDays { get; set; } = null;
+        public decimal? UsedDays { get; set; } = null;
     }
 }
